+++ src/main/java/uk/ac/gla/terrier/probos/JobUtils.java
@@ -190,8 +190,9 @@ public class JobUtils {
 		if (cmd.hasOption('N'))
 		{
 			String job_Name = cmd.getOptionValue('N');
-			if (job_Name.length() > 15)
-				job_Name = job_Name.substring(0, 15);
+			//Job names and outputs files are truncated #10
+			//if (job_Name.length() > 15)
+			//	job_Name = job_Name.substring(0, 15);
 			job.setJob_Name(job_Name);
 		}
 		
@@ -363,7 +364,7 @@ public class JobUtils {
 		//resolve to hostnames if necessary
 		if (job.getOutput_Path() == null)
 		{
-			job.setOutput_Path(hostname + ":" + cwd + File.separatorChar + job_Name + ".o");
+			job.setOutput_Path(hostname + ":" + cwd + File.separatorChar + job_Name + ".o" + "${PBS_JOBID}");
 		}
 		else
 		{
@@ -371,7 +372,7 @@ public class JobUtils {
 		}
 		if (job.getError_Path() == null)
 		{
-			job.setError_Path(hostname + ":" + cwd + File.separatorChar + job_Name + ".e");
+			job.setError_Path(hostname + ":" + cwd + File.separatorChar + job_Name + ".e" + "${PBS_JOBID}");
 		}
 		else
 		{